+++ packages/stateful-components/src/providerIcons.tsx
@@ -5,6 +5,7 @@
 import { KeplrIcon } from "@evmosapps/icons/KeplrIcon";
 import { LeapIcon } from "@evmosapps/icons/LeapWalletIcon";
 import { MetamaskIcon } from "@evmosapps/icons/MetamaskIcon";
+import { RabbyIcon } from "@evmosapps/icons/RabbyWalletIcon";
 import { WalletConnectIcon } from "@evmosapps/icons/WalletConnectIcon";
 
 export const ProvidersIcons: Record<
@@ -15,4 +16,5 @@ export const ProvidersIcons: Record<
   Keplr: KeplrIcon,
   WalletConnect: WalletConnectIcon,
   Leap: LeapIcon,
+  "Rabby Wallet": RabbyIcon,
 };